requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]
bcrypt
langgraph
langchain
langchain-anthropic
tavily-python
reportlab
